common: edk2: tools: use ccache
Set BUILD_CC and BUILD_CXX to include $(CCACHE) in order to improve the
build speed of the EDK2 tools. These variables default to gcc and g++
respectively so keep these commands but simply prepend the ccache
commmand.

This make the tools build time drop from 9 seconds to 1 second on my
laptop (when the cache is primed).

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>