common: edk2: use ccache
Speed up the EDK2 build further by enabling ccache. Contrary to many
projects, prepending the ccache command to the cross-compiler prefix
doesn't work, because the EDK2 build script passes the string to the
shell in double quotes. As a result the command is something like:

 "/usr/bin/ccache /path/to/aarch64-linux-gnu-" -g -Os ... <file.c>

which results in a "command not found" error.

The trick is to create a wrapper script for each cross-compile tool
that the build requires: <prefix>gcc, <prefix>gcc-ar and
<prefix>objcopy. This is easily done just before running the make
command in the EDK2 source directory.

With this commit (and the previous ones), the EDK2 build time is
reduced from 1m21s to 26s on my laptop (with a primed cache).

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>